konexion_backend/pyproject.toml
[project]
name = "web-konexion"
version = "0.1.0"
description = "Inference Web Models"
requires-python = ">=3.12, <3.14.0"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "loguru>=0.7.3",
    "mcp>=1.26.0",
    "openai>=2.30.0",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "websockets>=15.0.1",
    "authlib>=1.5.0",
    "python-jose[cryptography]>=3.5.0",
    "aiosqlite>=0.20.0",
    "slowapi>=0.1.9",
    "httpx>=0.28.0",
    "itsdangerous>=2.2.0",
    "greenlet>=3.2.4",
]

[dependency-groups]
dev = [
    "bandit[toml]>=1.8.6",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "ruff>=0.12.12",
    "yamllint>=1.37.1",
]

[tool.uv]
package = false
